Add K2000 Animation

+++ lightstep_fw/animations/K2000.h
@@ -1,5 +1,5 @@
 #include "../luz.h"
 #include <FastLED.h>
 
-void K2000(CRGB *leds, int time);
+void K2000(CRGB *leds, uint8_t time);
